requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
torch>=1.10.0
torchaudio>=0.10.0
librosa>=0.8.1
matplotlib>=3.5.0
tqdm>=4.62.0

# Audio processing
pydub>=0.25.1
soundfile>=0.10.3
webrtcvad>=2.0.10

# Speaker diarization
pyannote.audio>=2.0.0
resemblyzer>=0.1.0

# ASR (speech recognition)
transformers>=4.18.0
whisper>=1.0.0

# Utilities
pyyaml>=6.0
pandas>=1.3.0
notebook>=6.4.0
